[FEAT] Activity Indicator 추가 구현

+++ KoreaOneStep/ViewModels/SettingViewModel.swift
@@ -11,11 +11,23 @@ final class SettingViewModel {
     
     let inputBookmarkRemoveAllCellTapTrigger: Observable<Void?> = Observable(nil)
     
+    let outputRemoveAllBookmarksToastMessage: Observable<String?> = Observable(nil)
+    
     init() {
-        inputBookmarkRemoveAllCellTapTrigger.bind { trigger in
+        inputBookmarkRemoveAllCellTapTrigger.bind { [weak self] trigger in
+            guard let weakSelf = self else { return }
+            
             guard let trigger = trigger else { return }
             
-            RealmManager.shared.deleteAll()
+            let bookmarkList: [Bookmark] = RealmManager.shared.read(Bookmark.self).map { $0 }
+            
+            if bookmarkList.count >= 1 {
+                let toastMessage = RealmManager.shared.deleteAll()
+                weakSelf.outputRemoveAllBookmarksToastMessage.value = toastMessage
+                return
+            }
+            
+            weakSelf.outputRemoveAllBookmarksToastMessage.value = nil
         }
     }
 }
